requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langchain_google_genai
sse_starlette
fastapi
uvicorn
langserve
